11/24/24 - Map swap

+++ code/game/objects/crystal_key.dm
@@ -1,3 +1,8 @@
+//RS FILE
+
+#define DELETE_OBSTICAL 1
+#define TOGGLE_OBSTICAL 2
+
 var/global/list/event_obstical_keys = list()
 
 /obj/event_key
@@ -7,7 +12,13 @@ var/global/list/event_obstical_keys = list()
 	anchored = TRUE
 	var/id
 	var/spent = FALSE
+	var/spent_state = "gold_tri"
 	var/mob/living/link
+	var/reusable = FALSE
+	var/closed_state = "crystal_key_spent"
+
+/obj/event_key/reusable
+	reusable = TRUE
 
 /obj/event_key/Initialize()
 	. = ..()
@@ -53,9 +64,17 @@ var/global/list/event_obstical_keys = list()
 
 /obj/event_key/proc/post_trigger()
 	var/turf/ourturf = get_turf(src)
-	ourturf.visible_message("<span class = 'warning'>\The [src] crumbles to dust!!!</span>",runemessage = "crumble crumble")
-	unregister_mob()
-	qdel(src)
+	if(reusable)
+		ourturf.visible_message("<span class = 'warning'>\The [src] rumbles as something moves in the distance!!!</span>",runemessage = "rumble rumble")
+		spent = FALSE
+	else
+		unregister_mob()
+		if(closed_state)
+			ourturf.visible_message("<span class = 'warning'>\The [src] shimmers as it closes up!!!</span>",runemessage = "clink")
+			icon_state = closed_state
+		else
+			ourturf.visible_message("<span class = 'warning'>\The [src] crumbles to dust!!!</span>",runemessage = "crumble crumble")
+			qdel(src)
 
 /obj/event_key/proc/seek_link()
 	for(var/mob/living/thing in get_turf(src))
@@ -74,6 +93,15 @@ var/global/list/event_obstical_keys = list()
 		UnregisterSignal(link, COMSIG_MOB_DEATH)
 		UnregisterSignal(link, COMSIG_PARENT_QDELETING)
 
+/obj/event_key/hitby(atom/movable/AM)
+	. = ..()
+	if(isobj(AM))
+		trigger()
+
+/obj/event_key/bullet_act(obj/item/projectile/P, def_zone)
+	. = ..()
+	trigger()
+
 /obj/event_obstical
 	name = "impassable rock"
 	desc = "A shiny, impassable rock!"
@@ -85,10 +113,24 @@ var/global/list/event_obstical_keys = list()
 	opacity = TRUE
 	pixel_x = -16
 	pixel_y = -16
+	var/trigger_mode = DELETE_OBSTICAL
+	var/closed_state = null
+	var/open_state = null
 
 /obj/event_obstical/Initialize()
 	. = ..()
 	global.event_obstical_keys += src
+
+	if(trigger_mode == TOGGLE_OBSTICAL)
+		opacity = density
+		if(density)
+			if(closed_state)
+				icon_state = closed_state
+			else
+				closed_state = icon_state
+		else
+			if(open_state)
+				icon_state = open_state
 
 /obj/event_obstical/Destroy()
 	global.event_obstical_keys -= src
@@ -96,8 +138,20 @@ var/global/list/event_obstical_keys = list()
 
 /obj/event_obstical/proc/post_trigger()
 	var/turf/ourturf = get_turf(src)
-	ourturf.visible_message("<span class = 'warning'>\The [src] crumbles to dust!!!</span>",runemessage = "crumble crumble")
-	qdel(src)
+	switch(trigger_mode)
+		if(DELETE_OBSTICAL)
+			ourturf.visible_message("<span class = 'warning'>\The [src] crumbles to dust!!!</span>",runemessage = "crumble crumble")
+			qdel(src)
+		if(TOGGLE_OBSTICAL)
+			if(!closed_state)
+				closed_state = icon_state
+			density = !density
+			opacity = density
+			if(density)
+				icon_state = closed_state
+			else
+				icon_state = open_state
+			ourturf.visible_message("<span class = 'warning'>\The [src] rumbles as it moves!!!</span>",runemessage = "rumble rumble")
 
 /obj/event_obstical/disguised
 	name = "wall"
@@ -124,7 +178,17 @@ var/global/list/event_obstical_keys = list()
 	name = "decorated pillar"
 	icon_state = "crystal_pillar"
 	opacity = FALSE
+	closed_state = "crystal_pillar"
+	open_state = "crystal_pillar_lowered"
+
+/obj/event_obstical/disguised/pillar/toggle
+	trigger_mode = TOGGLE_OBSTICAL
 
 /obj/event_obstical/disguised/obstical
 	name = "decorated wall"
 	icon_state = "crystal_obstical"
+	closed_state = "crystal_obstical"
+	open_state = "crystal_obstical_lowered"
+
+/obj/event_obstical/disguised/obstical/toggle
+	trigger_mode = TOGGLE_OBSTICAL
